core_edge/requirements.txt
# ─── Core trading / exchange ─────────────────────────────────────────────────
hyperliquid-python-sdk>=0.7.0

# ─── Data & analysis ─────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0

# ─── Dashboard ───────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ─── Ethereum / wallet ───────────────────────────────────────────────────────
eth-account>=0.11.0

# ─── Environment / config ────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── HTTP (transitive, pinned for stability) ─────────────────────────────────
requests>=2.31.0
